chore(@clayui/breadcrumb): update Breadcrumb component stories
